+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## [1.73.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.73.1...v1.73.2) (2024-10-04)
+
+
+### Bug Fixes
+
+* **searchresults:** L3-4061 matching text was not highlighting ([#357](https://github.com/PhillipsAuctionHouse/seldon/issues/357)) ([8ff6379](https://github.com/PhillipsAuctionHouse/seldon/commit/8ff6379bb51e0a5d9560561549b256d3265e4588))
+
+## [1.73.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.73.0...v1.73.1) (2024-10-04)
+
+
+### Bug Fixes
+
+* **accordion:** L3-3154 accordion item style was overriding <strong tags in CMS content ([#356](https://github.com/PhillipsAuctionHouse/seldon/issues/356)) ([3d47512](https://github.com/PhillipsAuctionHouse/seldon/commit/3d475123de1a62c017ae012efbe9f04c46b64ba9))
+
+# [1.73.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.72.2...v1.73.0) (2024-10-03)
+
+
+### Features
+
+* **detail:** L3-4057 detail list improvement ([#355](https://github.com/PhillipsAuctionHouse/seldon/issues/355)) ([4e1859a](https://github.com/PhillipsAuctionHouse/seldon/commit/4e1859a71b0564acff42c4f532b6219b40b6ce6b))
+
+## [1.72.2](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.72.1...v1.72.2) (2024-10-03)
+
+
+### Bug Fixes
+
+* **lotdetails:** L3-3145 dont use index key on carousel dots ([#354](https://github.com/PhillipsAuctionHouse/seldon/issues/354)) ([b66d1b7](https://github.com/PhillipsAuctionHouse/seldon/commit/b66d1b71c680de4cd751c7e321b02cef26612d54))
+
+## [1.72.1](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.72.0...v1.72.1) (2024-10-03)
+
+
+### Bug Fixes
+
+* **accordion:** L3-3154 accordion label and icon style fixes ([#352](https://github.com/PhillipsAuctionHouse/seldon/issues/352)) ([9b8727d](https://github.com/PhillipsAuctionHouse/seldon/commit/9b8727dce707205a0ef8a574efe4c209fd40a347))
+
+# [1.72.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.71.0...v1.72.0) (2024-10-02)
+
+
+### Features
+
+* **Detail:** L3-3155 create Detail component ([#335](https://github.com/PhillipsAuctionHouse/seldon/issues/335)) ([48eb283](https://github.com/PhillipsAuctionHouse/seldon/commit/48eb283dc993966ff82bfb9d032f83a5534f9ea6))
+
 # [1.71.0](https://github.com/PhillipsAuctionHouse/seldon/compare/v1.70.0...v1.71.0) (2024-10-02)